remove trigger components in favor of element slots
