--- packages/react-components/react-combobox/src/ComboboxBase.ts
@@ -1,1 +1,0 @@
-export * from './components/ComboboxBase/index';
