--- packages/react-components/react-combobox/src/ComboboxInput.ts
@@ -1,1 +1,0 @@
-export * from './components/ComboboxInput/index';
