--- packages/react-components/react-combobox/src/components/ComboButton/index.ts
@@ -1,5 +1,0 @@
-export * from './ComboButton';
-export * from './ComboButton.types';
-export * from './renderComboButton';
-export * from './useComboButton';
-export * from './useComboButtonStyles';
