+++ packages/react-components/react-combobox/src/components/Combobox/useCombobox.tsx
@@ -1,34 +1,36 @@
 import * as React from 'react';
+import { ChevronDownRegular as ChevronDownIcon } from '@fluentui/react-icons';
 import { getPartitionedNativeProps, resolveShorthand } from '@fluentui/react-utilities';
 import { useComboboxBaseSlots } from '../../ComboboxBase/useComboboxBaseSlots';
 import { useComboboxBaseState } from '../../ComboboxBase/useComboboxBaseState';
 import { useComboboxPopup } from '../../utils/useComboboxPopup';
 import { Listbox } from '../Listbox/Listbox';
-import type { DropdownProps, DropdownState } from './Dropdown.types';
+import type { ComboboxProps, ComboboxState } from './Combobox.types';
 
 /**
  * Create the state required to render Combobox.
  *
  * The returned state can be modified with hooks such as useComboboxStyles_unstable,
- * before being passed to renderDropdown_unstable.
+ * before being passed to renderCombobox_unstable.
  *
- * @param props - props from this instance of Dropdown
- * @param ref - reference to root HTMLElement of Dropdown
+ * @param props - props from this instance of Combobox
+ * @param ref - reference to root HTMLElement of Combobox
  */
-export const useDropdown_unstable = (props: DropdownProps, ref: React.Ref<HTMLButtonElement>): DropdownState => {
+export const useCombobox_unstable = (props: ComboboxProps, ref: React.Ref<HTMLInputElement>): ComboboxState => {
   const baseState = useComboboxBaseState(props);
 
   const { primary: triggerNativeProps, root: rootNativeProps } = getPartitionedNativeProps({
     props,
-    primarySlotTagName: 'button',
-    excludedPropNames: ['children'],
+    primarySlotTagName: 'input',
+    excludedPropNames: ['children', 'size'],
   });
 
-  const state: DropdownState = {
+  const state: ComboboxState = {
     components: {
       root: 'div',
+      expandIcon: 'span',
+      input: 'input',
       listbox: Listbox,
-      button: 'button',
     },
     root: resolveShorthand(props.root, {
       required: true,
@@ -37,12 +39,17 @@ export const useDropdown_unstable = (props: DropdownProps, ref: React.Ref<HTMLBu
         ...rootNativeProps,
       },
     }),
-    listbox: resolveShorthand(props.listbox, {
-      required: true,
+    expandIcon: resolveShorthand(props.expandIcon, {
+      defaultProps: {
+        children: <ChevronDownIcon />,
+      },
     }),
-    button: resolveShorthand(props.button, {
+    input: resolveShorthand(props.input, {
       required: true,
       defaultProps: triggerNativeProps,
+    }),
+    listbox: resolveShorthand(props.listbox, {
+      required: true,
     }),
     ...baseState,
   };
