--- packages/react-components/react-combobox/src/components/ComboboxInput/index.ts
@@ -1,5 +1,0 @@
-export * from './ComboboxInput';
-export * from './ComboboxInput.types';
-export * from './renderComboboxInput';
-export * from './useComboboxInput';
-export * from './useComboboxInputStyles';
