+++ packages/react-components/react-combobox/src/components/Dropdown/Dropdown.types.ts
@@ -6,11 +6,14 @@ export type DropdownSlots = {
   /* The root combobox slot */
   root: NonNullable<Slot<'div'>>;
 
-  /* The dropdown listbox slot */
-  listbox: NonNullable<Slot<typeof Listbox>>;
-
   /* The primary slot, the element with role="combobox" */
   button: NonNullable<Slot<'button'>>;
+
+  /* The dropdown arrow icon */
+  expandIcon: Slot<'span'>;
+
+  /* The dropdown listbox slot */
+  listbox: NonNullable<Slot<typeof Listbox>>;
 };
 
 /**
