+++ packages/react-components/react-combobox/src/components/Dropdown/useDropdown.tsx
@@ -1,34 +1,36 @@
 import * as React from 'react';
+import { ChevronDownRegular as ChevronDownIcon } from '@fluentui/react-icons';
 import { getPartitionedNativeProps, resolveShorthand } from '@fluentui/react-utilities';
 import { useComboboxBaseSlots } from '../../ComboboxBase/useComboboxBaseSlots';
 import { useComboboxBaseState } from '../../ComboboxBase/useComboboxBaseState';
 import { useComboboxPopup } from '../../utils/useComboboxPopup';
 import { Listbox } from '../Listbox/Listbox';
-import type { ComboboxProps, ComboboxState } from './Combobox.types';
+import type { DropdownProps, DropdownState } from './Dropdown.types';
 
 /**
  * Create the state required to render Combobox.
  *
  * The returned state can be modified with hooks such as useComboboxStyles_unstable,
- * before being passed to renderCombobox_unstable.
+ * before being passed to renderDropdown_unstable.
  *
- * @param props - props from this instance of Combobox
- * @param ref - reference to root HTMLElement of Combobox
+ * @param props - props from this instance of Dropdown
+ * @param ref - reference to root HTMLElement of Dropdown
  */
-export const useCombobox_unstable = (props: ComboboxProps, ref: React.Ref<HTMLInputElement>): ComboboxState => {
+export const useDropdown_unstable = (props: DropdownProps, ref: React.Ref<HTMLButtonElement>): DropdownState => {
   const baseState = useComboboxBaseState(props);
 
   const { primary: triggerNativeProps, root: rootNativeProps } = getPartitionedNativeProps({
     props,
-    primarySlotTagName: 'input',
-    excludedPropNames: ['children', 'size'],
+    primarySlotTagName: 'button',
+    excludedPropNames: ['children'],
   });
 
-  const state: ComboboxState = {
+  const state: DropdownState = {
     components: {
       root: 'div',
+      button: 'button',
+      expandIcon: 'span',
       listbox: Listbox,
-      input: 'input',
     },
     root: resolveShorthand(props.root, {
       required: true,
@@ -37,12 +39,17 @@ export const useCombobox_unstable = (props: ComboboxProps, ref: React.Ref<HTMLIn
         ...rootNativeProps,
       },
     }),
-    listbox: resolveShorthand(props.listbox, {
-      required: true,
-    }),
-    input: resolveShorthand(props.input, {
+    button: resolveShorthand(props.button, {
       required: true,
       defaultProps: triggerNativeProps,
+    }),
+    expandIcon: resolveShorthand(props.expandIcon, {
+      defaultProps: {
+        children: <ChevronDownIcon />,
+      },
+    }),
+    listbox: resolveShorthand(props.listbox, {
+      required: true,
     }),
     ...baseState,
   };
