+++ packages/react-components/react-combobox/src/stories/Combobox.stories.tsx
@@ -6,7 +6,6 @@ import bestPracticesMd from './ComboboxBestPractices.md';
 
 export { Default } from './ComboboxDefault.stories';
 export { CustomOptions } from './ComboboxCustomOptions.stories';
-export { Editable } from './ComboboxEditable.stories';
 export { Multiselect } from './ComboboxMultiselect.stories';
 export { Grouped } from './ComboboxGrouped.stories';
 export { Appearance } from './ComboboxAppearance.stories';
